requirements.txt
fastapi
uvicorn
python-telegram-bot
httpx
flask
flask-cors
pydantic
